+++ rom-name-finder/frontend/src/utils/constants.ts
@@ -2,3 +2,5 @@ export const ERR_SEARCH_TERM_EMPTY = 'Please input at least one arcade game name
 export const ERR_SEARCH_TERM_SHORT = 'Names should have at least 3 characters';
 export const ERR_DB_SELECT = 'Please select a ROMset';
 export const ERR_UNKNOWN = 'An unexpected error occurred';
+
+export const MAX_FILE_SIZE = 5 * 1024; // KB